+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 ## 0.1.12
+- Restore previous workflow context after finished processing
 - Update Cadence testing to support synchronous child workflows.
 
 ## 0.1.11